+++ labextension/src/components/cell-metadata/CellMetadataEditor.tsx
@@ -16,7 +16,7 @@
 
 import * as React from 'react';
 import { Notebook, NotebookPanel } from '@jupyterlab/notebook';
-import { MaterialSelect, MaterialSelectMulti } from '../Components';
+import { MaterialSelectMulti } from '../Components';
 import TagsUtils from '../../utils/TagsUtils';
 import { isCodeCellModel } from '@jupyterlab/cells';
 import CloseIcon from '@material-ui/icons/Close';
@@ -25,6 +25,7 @@ import { CellMetadataContext } from './CellMetadataContext';
 import { Button, IconButton } from '@material-ui/core';
 import { CellMetadataEditorDialog } from './CellMetadataEditorDialog';
 import { Input } from '../Input';
+import { Select } from '../Select';
 
 const CELL_TYPES = [
   { value: 'imports', label: 'Imports' },
@@ -365,7 +366,7 @@ export class CellMetadataEditor extends React.Component<IProps, IState> {
               }
               style={{ borderLeft: `2px solid ${cellColor}` }}
             >
-              <MaterialSelect
+              <Select
                 updateValue={this.updateCurrentCellType}
                 values={CELL_TYPES}
                 value={cellType}
